pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["Rajat Jacob <rajatjacob@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
transformers = "^4.37.2"
chromadb = "^0.4.22"
python-dotenv = "^1.0.1"
google-generativeai = "^0.3.2"
streamlit = "^1.31.0"
redis = {extras = ["hiredis"], version = "^5.0.1"}
nltk = "^3.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
